Moved test_integration from the folder 'react' to the folder 'core' because of wired type errors
