+++ examples/react-typescript/src/core/index.ts
@@ -4,8 +4,8 @@ export const App = new Agile({
     logJobs: true,
 });
 
-export const MY_STATE = App.State<string>("MyState", "my-state").persist();
-export const MY_STATE_2 = App.State<string>("MyState2").persist("my-state2");
+export const MY_STATE = App.State<string>("MyState", "my-state");//.persist();
+export const MY_STATE_2 = App.State<string>("MyState2");//.persist("my-state2");
 
 MY_STATE.watch("test", (value) => {
     console.log("Watch " + value);
