+++ packages/core/tests/collection/functions/update.function.spec.ts
@@ -1,7 +1,7 @@
 import 'mocha';
 import {expect} from 'chai';
 import {Agile, Group} from "../../../src";
-import {useAgile_Test} from "../../../../react/tests";
+import {useAgile_Test} from "../../test_integration";
 
 describe('Update Function Tests', () => {
     let rerenderCount = 0;
