+++ packages/core/tests/collection/group/functions/remove.function.spec.ts
@@ -1,7 +1,7 @@
 import 'mocha';
 import {expect} from "chai";
 import {Agile, Group} from "../../../../src";
-import {useAgile_Test} from "../../../../../react/tests";
+import {useAgile_Test} from "../../../test_integration";
 
 describe('Add function Tests', () => {
     let rerenderCount = 0;
