+++ packages/core/tests/computed/default.spec.ts
@@ -1,7 +1,7 @@
 import 'mocha';
 import {expect} from 'chai';
 import {Agile, Computed} from "../../src";
-import {useAgile_Test} from "../../../react/tests";
+import {useAgile_Test} from "../test_integration";
 
 describe('Default Computed Tests', () => {
     // Define Agile
