+++ packages/core/tests/state/functions/undo.function.spec.ts
@@ -1,7 +1,7 @@
 import 'mocha';
 import {expect} from 'chai';
 import {Agile} from "../../../src";
-import {useAgile_Test} from "../../../../react/tests";
+import {useAgile_Test} from "../../test_integration";
 
 describe('Undo Function Tests', () => {
     let rerenderCount = 0;
