+++ packages/core/tests/test_integration/ReadMe.md
@@ -1,0 +1,1 @@
+Has to be in this package otherwise wired type errors will occur
