+++ packages/core/tests/test_integration/hooks/useEvent_Test.ts
@@ -1,4 +1,4 @@
-import {Event, EventCallbackFunction} from '@agile-ts/core';
+import {Event, EventCallbackFunction} from '../../../src';
 
 export function useEvent_Test<E extends Event>(event: E, callback: EventCallbackFunction<E['payload']>) {
     event.on(callback);
